requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
pillow
statistics
svgwrite
math
base64
struct
zlib
csv
ShopifyAPI
oauth2client
google-api-python-client
sqlalchemy
psycopg2-binary
alembic
grpcio
grpcio-tools
bcrypt
pyjwt